Added history selection, transitions
Added a select box which allows the user to pick a timeframe for the
line chart. Options are 1 day, 3 days, 7 days, or 30 days. Also added a
transition that animates both the valueline and axes as they adapt to
the new timeframe. Also added a function to perform date operations and
conversions to adapt to Google Sheet's query needs. Makes use of
tabletop's query feature.